extract ImageViewHolder from Image model
