+++ ui/tract-renderer/src/main/java/org/cru/godtools/tract/model/Image.java
@@ -2,21 +2,13 @@ package org.cru.godtools.tract.model;
 
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
-import android.support.annotation.UiThread;
 import android.support.annotation.WorkerThread;
-import android.view.ViewGroup;
 
-import org.ccci.gto.android.common.picasso.view.PicassoImageView;
 import org.ccci.gto.android.common.util.XmlPullParserUtils;
-import org.cru.godtools.tract.R;
-import org.cru.godtools.tract.R2;
-import org.cru.godtools.tract.viewmodel.ResourceViewUtils;
 import org.xmlpull.v1.XmlPullParser;
 import org.xmlpull.v1.XmlPullParserException;
 
 import java.io.IOException;
-
-import butterknife.BindView;
 
 import static org.cru.godtools.tract.Constants.XMLNS_CONTENT;
 
@@ -54,21 +46,5 @@ public final class Image extends Content {
     void parseAttrs(@NonNull final XmlPullParser parser) {
         super.parseAttrs(parser);
         mResource = parser.getAttributeValue(null, XML_RESOURCE);
-    }
-
-    @UiThread
-    static final class ImageViewHolder extends BaseViewHolder<Image> {
-        @BindView(R2.id.image)
-        PicassoImageView mImage;
-
-        ImageViewHolder(@NonNull final ViewGroup parent, @Nullable final BaseViewHolder parentViewHolder) {
-            super(Image.class, parent, R.layout.tract_content_image, parentViewHolder);
-        }
-
-        @Override
-        void onBind() {
-            super.onBind();
-            ResourceViewUtils.bind(getResource(mModel), mImage);
-        }
     }
 }
